+++ src/app/page.tsx
@@ -1,0 +1,8 @@
+
+export default function Home() {
+  return (
+    <main className="flex min-h-screen flex-col items-center justify-between p-24">
+      Hi, This Application Provides Idea for Necessary Typescript for NextJS/ReactJS
+    </main>
+  )
+}
