+++ src/app/propsexample/page.tsx
@@ -1,6 +1,6 @@
-import PostList from "@/Utilities/Components/postList/PostList";
+import PostList from "@/Components/postList/PostList";
 
-
+PostList
 const PropsExample = ()=>{
     return (
         <PostList/>
